requirements.dev.txt
appdirs==1.4.4
attrs==19.3.0
autoflake==1.3.1
black==19.10b0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
google-api-core==1.17.0
google-auth==1.15.0
google-cloud-core==1.3.0
google-cloud-monitoring==0.36.0
google-cloud-trace==0.23.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
grpcio-reflection==1.29.0
grpcio-tools==1.29.0
idna==2.9
mypy==0.770
mypy-extensions==0.4.3
mypy-protobuf==1.21
opencensus==0.7.7
opencensus-context==0.1.1
opencensus-ext-grpc==0.7.1
opencensus-ext-stackdriver==0.7.2
opentelemetry-api==0.8b0
-e git+https://github.com/AndrewAXue/opentelemetry-python.git@7732d035f2d1e782eb4bc94efb2e1c15428aad3b#egg=opentelemetry_ext_cloud_trace&subdirectory=ext/opentelemetry-ext-cloud-trace
opentelemetry-ext-grpc==0.8b0
opentelemetry-sdk==0.8b0
pathspec==0.8.0
pkg-resources==0.0.0
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyflakes==2.2.0
pytz==2020.1
regex==2020.5.14
requests==2.23.0
rsa==4.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
